add commnet count

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable("id")Integer id,  Model model) {
+    public String publish(@PathVariable("id")Long id,  Model model) {
 
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -46,7 +46,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam(value = "id" ,required = false) Integer id,
+            @RequestParam(value = "id" ,required = false) Long id,
             HttpServletRequest request,
             Model model)
     {
